+++ src/main/kotlin/no/uutilsynet/testlab2testing/aggregering/AggregeringPerTestregelDTO.kt
@@ -13,7 +13,7 @@ data class AggregeringPerTestregelDTO(
     val talSiderSamsvar: Int,
     val talSiderBrot: Int,
     val talSiderIkkjeForekomst: Int,
-    val testregelGjennomsnittlegSideSamsvarProsent: Float?,
-    val testregelGjennomsnittlegSideBrotProsent: Float?,
+    val testregelGjennomsnittlegSideSamsvarProsent: Double?,
+    val testregelGjennomsnittlegSideBrotProsent: Double?,
     val testgrunnlagId: Int?
 )
